chore: rename grammar -> parser
This is what it is. A parser (Lezer) for feelers.
