+++ lib/index.js
@@ -1,4 +1,4 @@
 export { evaluate } from './interpreter/index.js';
 export { FeelersEditor } from './editor/index.js';
 
-export * from './grammar/index.js';
+export * from './parser/index.js';
