+++ lib/interpreter/index.js
@@ -1,4 +1,4 @@
-import { parser, buildSimpleTree } from '../grammar//index.js';
+import { parser, buildSimpleTree } from '../parser/index.js';
 import { evaluate as evaluateFeel } from '@bpmn-io/feelin';
 
 /**
