+++ lib/parser/highlight.js
