+++ lib/parser/tokens.js
