+++ package.json
@@ -7,7 +7,7 @@
   "scripts": {
     "all": "run-s build lint test",
     "test": "karma start karma.conf.cjs",
-    "generate:parser": "lezer-generator lib/grammar/feelers.grammar -o lib/grammar/parser.js",
+    "generate:parser": "lezer-generator lib/parser/feelers.grammar -o lib/parser/parser.js",
     "bundle": "rollup -c",
     "dev": "npm test -- --auto-watch --no-single-run",
     "build": "npm run generate:parser && npm run bundle",
