+++ test/spec/playground.spec.js
@@ -3,8 +3,8 @@ import { marked } from 'marked';
 
 import { initialContext, initialTemplate } from '../testData.js';
 
-import { FeelersEditor } from '../../src/editor/index.js';
-import { evaluate } from '../../src/interpreter/index.js';
+import { FeelersEditor } from '../../lib/editor/index.js';
+import { evaluate } from '../../lib/interpreter/index.js';
 
 
 const singleStart = window.__env__ && window.__env__.SINGLE_START;
